+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 ## [Unreleased]
 
+## 1.0.3 - 
+
+Added `letter-spacing` as preferred `ls` shortname
+Fixed unsetting props when using setter functions
+Use getComputedStyle for property registration (fixes safari 5)
+
+## 1.0.2 - 
+
+Convenience toString helper changed to valueOf
+Fixed px being added to flex shorthand unexpectedly
+Lazy registration of px value properties
+Add multiple properties per line for css strings
+
+## 1.0.1 - 
+
+Fixed px property registration in edge
+Fixed px addition for shorthands
+
 ## 1.0.0 - 2017-09-05
 
 First stable release of `bss`. Changes from here on will follow [semver](http://semver.org/).